Added token based accessing limits

+++ src/main/java/comp307/backend/booking/Service/BookingService.java
@@ -32,8 +32,8 @@ public class BookingService {
     //Type 3
     //startDateTimes/endDateTimes are not to be confused with listing all the weeks. Instead, they list the first weeks time slots. For example Monday Jan 1st 2/3 pm and Tuesday Jan 2nd 3/4 pm. 
     //From this we use weeksToRepeat to loop and make booking slots for the future weeks.
-    public void createRecurringBookingSlot(String ownerEmail, List<LocalDateTime> startDateTimes, List<LocalDateTime> endDateTimes, int weeksToRepeat) {
-        User owner = userRepository.findById(ownerEmail).orElseThrow(() -> new RuntimeException("User " + ownerEmail + " not found."));
+    public void createRecurringBookingSlot(String ownerToken, List<LocalDateTime> startDateTimes, List<LocalDateTime> endDateTimes, int weeksToRepeat) {
+        User owner = userRepository.findByaccessToken(ownerToken).orElseThrow(() -> new RuntimeException("User " + ownerToken + " not found."));
 
         if (startDateTimes.size() != endDateTimes.size()) {
             throw new RuntimeException("Size of start date times and end date times do not match.");
@@ -56,20 +56,28 @@ public class BookingService {
         throw new RuntimeException("Not implemented yet");
     }
 
-    public List<BookingSlot> getAllOwnedSlots(String ownerEmail) {
-        User owner = userRepository.findById(ownerEmail).orElseThrow(() -> new RuntimeException("User " + ownerEmail + " not found."));
+    public List<BookingSlot> getAllOwnedSlots(String ownerToken) {
+        User owner = userRepository.findByaccessToken(ownerToken).orElseThrow(() -> new RuntimeException("User " + ownerToken + " not found."));
 
         return bookingSlotRepository.findByOwner(owner);
         
     }
 
-    public List<BookingSlot> getAllAvailableOwnedSlots(String ownerEmail) {
-        return getAllOwnedSlots(ownerEmail).stream().filter(bookingSlot -> (bookingSlot.getSlotStatus() == BookingSlot.BookingSlotStatus.AVAILABLE)).toList();
+    public List<BookingSlot> getAllAvailableOwnedSlots(String ownerEmail, String userToken) {
+        userRepository.findByaccessToken(userToken).orElseThrow(() -> new RuntimeException("User " + userToken + " not found."));
+        User owner = userRepository.findById(ownerEmail).orElseThrow(() -> new RuntimeException("User " + ownerEmail + " not found."));
+        return getAllOwnedSlots(owner.getAccessToken()).stream().filter(bookingSlot -> (bookingSlot.getSlotStatus() == BookingSlot.BookingSlotStatus.AVAILABLE)).toList();
     }
 
     //maybe add email service, or frontend could default open email with all the people whose bookings got cancelled in which case can return list of emails that should be notified
-    public void cancelBookingSlot(Long bookingSlotId) {
+    public void cancelBookingSlot(String ownerToken, Long bookingSlotId) {
         BookingSlot bookingSlot = bookingSlotRepository.findById(bookingSlotId).orElseThrow(() -> new RuntimeException("Slot " + bookingSlotId + " not found."));
+        User owner = userRepository.findByaccessToken(ownerToken).orElseThrow(() -> new RuntimeException("Invalid Token"));
+
+        if (!bookingSlot.getOwner().equals(owner)) {
+            throw  new RuntimeException("Invalid access");
+        }
+
         bookingSlot.setSlotStatus(BookingSlot.BookingSlotStatus.CANCELLED);
         bookingSlotRepository.save(bookingSlot);
 
@@ -83,8 +91,8 @@ public class BookingService {
 
 
     //BOOKING
-    public Booking book(Long bookingSlotId, String reserveeEmail) {
-        User reservee = userRepository.findById(reserveeEmail).orElseThrow(() -> new RuntimeException("User " + reserveeEmail + " not found."));
+    public Booking book(Long bookingSlotId, String reserveeToken) {
+        User reservee = userRepository.findByaccessToken(reserveeToken).orElseThrow(() -> new RuntimeException("User " + reserveeToken + " not found."));
         BookingSlot bookingSlot = bookingSlotRepository.findById(bookingSlotId).orElseThrow(() -> new RuntimeException("Slot " + bookingSlotId + " not found."));
 
         if (!bookingSlot.getSlotStatus().equals(BookingSlot.BookingSlotStatus.AVAILABLE)) {
@@ -115,11 +123,17 @@ public class BookingService {
     }
 
     //whether its type 2 or type 3, either way the booking will become available when unbooked because it either had infinite space or now has at least 1 space
-    public void unbook(Long bookingId) {
+    public void unbook(Long bookingId, String reserveeToken) {
+        User reservee = userRepository.findByaccessToken(reserveeToken).orElseThrow(() -> new RuntimeException("User " + reserveeToken + " not found."));
         Booking booking = bookingRepository.findById(bookingId).orElseThrow(() -> new RuntimeException("Booking " + bookingId + " not found."));
-        BookingSlot bookingSlot = booking.getBookingSlot();
-        bookingSlot.setSlotStatus(BookingSlot.BookingSlotStatus.AVAILABLE);
-        bookingSlotRepository.save(bookingSlot);
-        bookingRepository.delete(booking);
+
+        if (booking.getReservee().equals(reservee)) {
+            BookingSlot bookingSlot = booking.getBookingSlot();
+            bookingSlot.setSlotStatus(BookingSlot.BookingSlotStatus.AVAILABLE);
+            bookingSlotRepository.save(bookingSlot);
+            bookingRepository.delete(booking);
+        } else {
+            throw new RuntimeException("Illegal Access");
+        }
     }   
 }
